web-scraping/studocu/requirements.txt
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.2
idna==3.10
pillow==11.3.0
reportlab==4.4.3
requests==2.32.4
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
